Task (#44)
* TRACK WORD LOADING

* change track words history keeping

* update limit type

* ordery by documentid

* success on track words

* perfect syncing feature

* refactoring wordnik manager

* full state for word resolution

* lots of improvement on word

* lots of improvement on word

* improve retry on not found resolution for word

* not found code 404

* improve definition object

* keep url with definition

* keep url with definition

+++ tools/src/main/kotlin/com/dreampany/tools/data/source/memory/provider/AppProvider.kt
@@ -20,7 +20,7 @@ class AppProvider
     val mapper: AppMapper
 ) : MediaProvider<App>() {
 
-    override fun getItems(limit: Int): List<App>? {
+    override fun getItems(limit: Long): List<App>? {
         val pm = AndroidUtil.getPackageManager(context)
         val infos = AndroidUtil.getInstalledApps(context, pm)
         val result = mutableListOf<App>()
